refactored code for better soc
