+++ packages/be-nest/src/entity/person.entity.ts
@@ -1,4 +1,4 @@
-import { BaseCreateable } from 'src/base_entity/base-creatable';
+import { BaseCreateable } from '@entity/base-creatable';
 import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity()
