--- packages/be-nest/src/genre/genre.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Genre } from './genre.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Genre])],
-  exports: [TypeOrmModule],
-})
-export class GenresModule {}
