+++ packages/be-nest/src/modules/genre.module.ts
@@ -1,0 +1,13 @@
+import { GenresController } from '@controller/genre.controller';
+import { Genre } from '@entity/genre.entity';
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { GenresService } from '@services/genre.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Genre])],
+  providers: [GenresService],
+  controllers: [GenresController],
+  exports: [TypeOrmModule],
+})
+export class GenreHttpModule {}
