+++ packages/be-nest/src/services/genre.service.ts
@@ -1,7 +1,7 @@
+import { Genre } from '@entity/genre.entity';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Genre } from './genre.entity';
 
 @Injectable()
 export class GenresService {
