+++ packages/be-nest/src/services/person.service.ts
@@ -1,7 +1,7 @@
+import { Person } from '@entity/person.entity';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Person } from './person.entity';
 
 @Injectable()
 export class PersonService {
